feat: removed mdns discovery from the integration
The dock will be added to the config when setting up for the first time. After this, new docks can be added with the web configurator. The web configurator will do the mdns discovery.

There is no need to run a scan for new docks on the network, as it's highly unlikely that new docks will pop up every minute or so :)